requirements.txt
mcp-server>=0.1.0
fastmcp>=0.1.0
mysql-connector-python>=8.0.0
boto3>=1.26.0
starlette>=0.27.0
python-dotenv>=1.0.0